Move CommitsBetweenBounds into analytics where it belongs
